requirements.txt
python-dotenv
ffmpeg-python
openai
yt-dlp
